requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
discord.py==2.6.3
dotenv==0.9.9
EorzeaEnv==2.2.11
frozenlist==1.7.0
h11==0.16.0
idna==3.10
multidict==6.6.4
numpy==2.3.3
outcome==1.3.0.post0
propcache==0.3.2
pycparser==2.23
PySocks==1.7.1
python-dotenv==1.1.1
rapidfuzz==2.15.2
requests==2.32.5
selenium==4.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
